store/requirements.txt
# requirements.txt

# Standard dependencies.
omegaconf

# FastAPI dependencies.
aiohttp
aiosmtplib
fastapi
pyjwt
python-multipart

# Deployment dependencies.
uvicorn[standard]
